+++ packages/cli/chain.config.js
@@ -1,0 +1,6 @@
+// [cli 配置]
+module.exports = ({ config }) => {
+  return () => {
+    console.log('来自cli的配置')
+  }
+}