+++ packages/cli/lib/util/tryGetNewerRange.js
@@ -1,4 +1,4 @@
-const { semver } = require('../@pkb/shared-utils')
+const { semver } = require('@pkb/shared-utils')
 
 const leadRE = /^(~|\^|>=?)/
 const rangeToVersion = r => r.replace(leadRE, '').replace(/x/g, '0')