+++ packages/packages-cli/cli/lib/util/registries.js
@@ -1,8 +1,0 @@
-const registries = {
-  npm: 'https://registry.npmjs.org',
-  yarn: 'https://registry.yarnpkg.com',
-  // taobao: 'https://registry.npm.taobao.org',
-  pnpm: 'https://registry.npmjs.org'
-}
-
-module.exports = registries