+++ api/PluginAPI.js
@@ -4,13 +4,6 @@ const path = require('path')
 exports.genCacheConfig = function (id, partialIdentifier, configFiles = []) {
   const fs = require('fs')
   const cacheDirectory = this.resolve(`node_modules/.cache/${id}`)
-  // replace \r\n to \n generate consistent hash
-  // const fmtFunc = conf => {
-  //   if (typeof conf === 'function') {
-  //     return conf.toString().replace(/\r\n?/g, '\n')
-  //   }
-  //   return conf
-  // }
 
   const variables = {
     partialIdentifier,
@@ -18,10 +11,7 @@ exports.genCacheConfig = function (id, partialIdentifier, configFiles = []) {
     'cache-loader': require('cache-loader/package.json').version,
     env: process.env.NODE_ENV,
     test: !!process.env.VUE_CLI_TEST,
-    config: [
-      // fmtFunc(this.service.projectOptions.chainWebpack),
-      // fmtFunc(this.service.projectOptions.configureWebpack)
-    ]
+    config: []
   }
 
   if (!Array.isArray(configFiles)) {