+++ bin/webpack-box.js
@@ -1,14 +1,13 @@
 #!/usr/bin/env node
-const { getAllCwds } = require('../api/CwdAPI')
-
 const {
   injectCommand,
+  getAllCommands,
   commandComplete
 } = require('../api/CommandAPI')
 const PluginAPI = require('../api/PluginAPI')
 
 // 注册命令行
-getAllCwds().forEach(cwd => {
+getAllCommands().forEach(cwd => {
   cwd({ injectCommand, api: PluginAPI })
 })
 // 命令行注册完成